[API USER] - Fix thay đổi thông tin cá nhân

+++ src/main/java/com/main/dto/UsersDto.java
@@ -4,6 +4,7 @@ import com.main.entity.Roles;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.sql.Timestamp;
 import java.util.Collection;
@@ -41,8 +42,8 @@ public class UsersDto implements Serializable {
 
     String token;
 
-    Integer totalOrderPrice;
+    BigDecimal totalOrderPrice;
 
-    Integer orderCount;
+    BigDecimal orderCount;
 
 }
